+++ functions/src/definitions/TanamPublishStatus.ts
@@ -1,0 +1,5 @@
+export enum TanamPublishStatus {
+  Published = "published",
+  Unpublished = "unpublished",
+  Scheduled = "scheduled",
+}
