+++ hosting/src/hooks/useGenkitArticle.tsx
@@ -2,7 +2,7 @@ import {TanamDocumentClient} from "@/models/TanamDocumentClient";
 import {UserNotification} from "@/models/UserNotification";
 import {firestore, storage} from "@/plugins/firebase";
 import {generateArticle} from "@/plugins/genkit/article";
-import {TanamPublishStatus} from "@functions/models/TanamDocument";
+import {TanamPublishStatus} from "@functions/definitions/TanamPublishStatus";
 import {ArticleSchema} from "@functions/schemas/article";
 import {collection, doc, getDocs, limit, orderBy, query, setDoc, where} from "firebase/firestore";
 import {getDownloadURL, ref, uploadBytes} from "firebase/storage";
